fixes #12
adds an argument for global out_of_slice_slice argument for calculating
a celltype specific config. This allows the saving of out_of_slice_slice
using python generation

+++ src/napari_3d_counter/celltype_config.py
@@ -52,7 +52,7 @@ class CellTypeConfig:  # pylint: disable=too-many-instance-attributes
     "the keyboard binding to switch to this celltype"
     outline_size: Optional[float] = None
     "The size of the circle around the cell"
-    out_of_slice_point_size: Optional[int] = None
+    out_of_slice_point_size: Optional[float] = None
     "The size of the out of slice point"
     face_color: Optional[MatplotlibColor] = None
     "The filled in color of the point"
